Restructure agents and integrate analyses
