+++ Agents/Legalagent/legalcheck.py
